pyproject.toml
[tool.poetry]
name = "aioli"
version = "0.5.1"
description = "Non-blocking Web API Framework and Toolkit"
authors = ["Robert Wikman <rbw@vault13.org>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/aioli-framework/aioli"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Code Generators",
]

[tool.poetry.dependencies]
python = "^3.6"
starlette = "^0.12.0b3"
uvloop = "^0.12.1"
uvicorn = "^0.6.1"
gunicorn = "^19.9.0"
marshmallow = "^3.0.0"
ujson = "^1.35"
poetry = "^0.12.17"
importlib_metadata = "^0.19.0"
texttable = "^1.6.2"
pyyaml = "^5.1.2"
click = "^7.0"
click_repl = "^0.1.6"

[tool.poetry.dev-dependencies]
twine = "^1.13"
sphinx = "^2.1.1"
pytest = "^4.4.0"
pytest-cov = "^2.7.1"
autoflake = "^1.3"
codecov = "^2.0.15"
docutils = { version = "0.15" }
black = { version = "^19.3b0", allows-prereleases = true }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
aioli = "aioli.cli:run"
